RemoveThatBG!/PythonBackend/requirements.txt
Flask>=3.0.0
waitress>=3.0.0
rembg[gpu]>=2.0.50,<2.0.62
Pillow>=9.0.0,<12.0.0
onnxruntime>=1.15.0
